fix(dataset,netcdf): tidy the last packing edges the review found
L3: the cube writer took band 1's `scale_factor` / `add_offset` and stamped them on
every variable, mislabelling a collection whose bands carry different factors. Each
per-band variable now gets its own recipe; the single 4-D `data` variable, which has
only one place to put one, takes it only when every band agrees.

L4: `stats` returned its four numbers through a `float32` frame. That predates this
work but now carries physical values, and a band packed at 1e-5 around an offset of
273.15 has more significant digits than `float32` holds -- rounding away exactly the
resolution the packing existed to preserve.

L5: `_stack_timesteps`' empty branch promised the collection's stored dtype while
the non-empty branch returns physical `float64`, so `head(0).dtype` and
`head(1).dtype` disagreed on a packed collection. Same reasoning as the rank guard
directly above it.

L6: restores the coverage the earlier rewrite lost. The scale-only / offset-only
cases moved to a raster, which exercises the band path; a `NetCDF` variable holding
a half-set pair in Python is the other resolver branch and is now covered directly.

The cube-writer stub is a real `Dataset` rather than a namespace, since the writer
asks the template for its packing as well as its axes -- modelling only the axes is
what let it drift.

+++ src/pyramids/dataset/engines/analysis.py
@@ -275,7 +275,12 @@ class Analysis(_Engine["Dataset"]):
             df = pd.DataFrame(
                 index=self._ds.band_names,
                 columns=["min", "max", "mean", "std"],
-                dtype=np.float32,
+                # `float64`, not `float32`: these are physical values now, and a band
+                # packed at 1e-5 around an offset of 273.15 -- an ordinary way to store
+                # temperature -- has more significant digits than `float32` carries.
+                # Rounding them here would throw away exactly the resolution the
+                # packing existed to preserve.
+                dtype=np.float64,
             )
             for i in range(self._ds.band_count):
                 if mask is not None and dst is not None:
@@ -286,7 +291,12 @@ class Analysis(_Engine["Dataset"]):
             df = pd.DataFrame(
                 index=[self._ds.band_names[band]],
                 columns=["min", "max", "mean", "std"],
-                dtype=np.float32,
+                # `float64`, not `float32`: these are physical values now, and a band
+                # packed at 1e-5 around an offset of 273.15 -- an ordinary way to store
+                # temperature -- has more significant digits than `float32` carries.
+                # Rounding them here would throw away exactly the resolution the
+                # packing existed to preserve.
+                dtype=np.float64,
             )
             if mask is not None and dst is not None:
                 df.iloc[0, :] = dst.analysis._get_stats(band, approx_ok=approx_ok)
